Create working chapters
For Flashcards, only show First Words section for now for all languages.
This code needs to be refactored to remove generator. The generator code
was not good because it made the useEffect non-idempotent. So getting
all words is much cleaner and there is no large downside. It's we'll
likely never need to worry about performance for getAllWords because the
wordlist is small and split by Chapter anyways.